requirements.txt
asgiref==3.2.3
blinker==1.4
canonicalwebteam.flask-base==0.3.3
canonicalwebteam.yaml-responses==1.1.1
certifi==2019.11.28
chardet==3.0.4
django-openid-auth==0.15
future==0.16.0
gunicorn==19.10.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
prometheus-client==0.4.2
pycparser==2.19
pymacaroons==0.13.0
python3-openid==3.1.0
requests==2.22.0
requests-cache==0.5.2
talisker==0.15.0
urllib3==1.25.7
